+++ .prettierrc
@@ -1,0 +1,6 @@
+{
+    "bracketSpacing": true,
+    "tabWidth": 2,
+    "semi": true,
+    "singleQuote": true
+}
