+++ src/app/common/title-bar/title-bar.component.html
@@ -1,0 +1,8 @@
+<ion-header>
+  <ion-toolbar>
+    <ion-buttons slot="start">
+      <ion-menu-button></ion-menu-button>
+    </ion-buttons>
+    <ion-title>{{ appTitleService.title$ | async }}</ion-title>
+  </ion-toolbar>
+</ion-header>
