--- src/app/home/home.page.scss
@@ -1,4 +1,0 @@
-.welcome-card ion-img {
-  max-height: 35vh;
-  overflow: hidden;
-}
