+++ src/app/shopper/components/shoplist-edition/shoplist-edition.component.html
@@ -3,6 +3,7 @@
     <ion-icon name="add-circle-outline" slot="start"></ion-icon>
     <ion-input (keyup.enter)="createShoplistItem(itemForm.value.name)" clearInput placeholder="New Item"
       formControlName="name" required></ion-input>
+    <ion-button color="primary" (click)="saveAsTemplate()">Save<ion-icon slot="start" name="albums"></ion-icon></ion-button>
   </ion-item>
   <div class="autocomplete-container" *ngIf="itemForm.value.name?.length > 2 && (products$ | async) as products">
     <ion-list>
