+++ src/app/shopper/shopper.page.html
@@ -1,12 +1,4 @@
-<ion-header>
-  <ion-toolbar>
-    <ion-buttons slot="start">
-      <ion-menu-button></ion-menu-button>
-    </ion-buttons>
-    <ion-title>{{ appTitleService.title$ | async }}</ion-title>
-  </ion-toolbar>
-</ion-header>
-
+<app-title-bar></app-title-bar>
 <ion-content>
   <ion-router-outlet></ion-router-outlet>
 </ion-content>
