+++ src/app/shopper/shopper.page.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AppTitleService } from './services/app-title.service';
+import { AppTitleService } from '../common/services/app-title.service';
 
 @Component({
   selector: 'app-shopper',
