+++ src/app/templates/templates-page/templates-page.component.html
@@ -1,8 +1,10 @@
-<div>
-  <p>Templates</p>
-  <ul>
-    <li *ngFor="let template of templates | async">
-      {{ template.label }} - {{ template.items.length || 0 }}
-    </li>
-  </ul>
-</div>
+<app-title-bar></app-title-bar>
+<ion-content>
+  <div>
+    <ul>
+      <li *ngFor="let template of templates | async">
+        {{ template.label }} - {{ template.items.length || 0 }}
+      </li>
+    </ul>
+  </div>
+</ion-content>
