+++ src/app/templates/templates.module.ts
@@ -3,6 +3,8 @@ import { CommonModule } from '@angular/common';
 
 import { TemplatesRoutingModule } from './templates-routing.module';
 import { TemplatesPageComponent } from './templates-page/templates-page.component';
+import { IonicModule } from '@ionic/angular';
+import { ChillModule } from '../common/chill.module';
 
 
 @NgModule({
@@ -11,6 +13,8 @@ import { TemplatesPageComponent } from './templates-page/templates-page.componen
   ],
   imports: [
     CommonModule,
+    IonicModule,
+    ChillModule,
     TemplatesRoutingModule
   ]
 })
